Improve build
