+++ core/babel-preset/index.js
@@ -5,5 +5,16 @@
 
 module.exports = {
   presets: ['env', 'react', 'es2015'],
-  plugins: ['transform-object-rest-spread', 'transform-class-properties']
+  plugins: [
+    'transform-object-rest-spread',
+    'transform-class-properties',
+    [
+      'babel-plugin-module-resolver',
+      {
+        alias: {
+          '@auth0/cosmos': './core/components'
+        }
+      }
+    ]
+  ]
 }
