+++ core/babel-preset/package.json
@@ -1,12 +1,13 @@
 {
   "name": "@auth0/babel-preset-cosmos",
   "description": "babel preset to use cosmos",
-  "version": "0.15.1",
+  "version": "0.15.2",
   "repository": "auth0/cosmos",
   "main": "index.js",
   "author": "siddharthkp",
   "license": "MIT",
   "dependencies": {
+    "babel-plugin-module-resolver": "^3.1.3",
     "babel-plugin-styled-components": "1.5.1",
     "babel-plugin-transform-class-properties": "6.24.1",
     "babel-plugin-transform-object-rest-spread": "6.26.0",
