+++ package.json
@@ -20,7 +20,8 @@
     "scripts": "nps -s",
     "i": "nps-i",
     "dev": "echo Our scripts have changed, run yarn scripts to know more",
-    "build": "nps -s production.build",
+    "build": "nps -s deploy.default",
+    "build-site": "nps -s production.build",
     "start": "nps -s production.start",
     "test": "nps -s test",
     "precommit": "nps icons.build && lint-staged",
